+++ src/page-modules/departures/nearest-stop-places/index.tsx
@@ -1,4 +1,4 @@
-import { MapWithHeader } from '@atb/modules/map';
+import { MapWithHeader } from '@atb/components/map';
 import { GeocoderFeature } from '../types';
 import {
   NearestStopPlacesData,
