+++ src/page-modules/departures/stop-place/index.tsx
@@ -1,4 +1,4 @@
-import { MapWithHeader } from '@atb/modules/map';
+import { MapWithHeader } from '@atb/components/map';
 import {
   DepartureData,
   Quay,
